Added patches to the repo and no longer auto-generate.  create-db-diffs will only generate db patch file if they don't exist, and filtering of unwanted diffs has been moved to the spec file.

+++ bin/create-db-diffs.rb
@@ -72,9 +72,11 @@ class DiffGenerator
         if count > 0
           patch = new.generate_patch(old)
           filename = "#{patch_dir}/db-#{ver}.wpatch"
-          puts "Generating patch file #{filename}"
-          File.open(filename, "w") do |of|
-            of.write(patch.to_yaml)
+          if !File.exist?(filename)
+            puts "Generating patch file #{filename}"
+            File.open(filename, "w") do |of|
+              of.write(patch.to_yaml)
+            end
           end
         end
         count += 1
